api/requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
GitPython==3.1.49
gunicorn==23.0.0
